tema3/adapter/requirements.txt
paho-mqtt==1.5.0
influxdb==5.2.3
